# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a server ecommerce.",
        "title": "Swagger Golang Ecommerce",
        "contact": {
            "name": "Nguyễn Đại Nghĩa",
            "url": "https://www.facebook.com/shy.leopard.beo",
            "email": "nghiabeo1605@gmail.com"
        }
    },
    "host": "localhost:8080",
    "paths": {
        "/api/auth/change-password": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "ChangePassword account",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "ChangePassword accounts",
                "parameters": [
                    {
                        "description": "ChangePassword user",
                        "name": "tags",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_NghiaLeopard_Go-Ecommerce-Backend_internal_api_handler_request.ChangePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Change password success",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/auth/forgot-password": {
            "post": {
                "description": "ForgotPassword account",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "ForgotPassword accounts",
                "parameters": [
                    {
                        "description": "ForgotPassword user",
                        "name": "tags",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_NghiaLeopard_Go-Ecommerce-Backend_internal_api_handler_request.ForgotPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Forgot password success",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "description": "Login account",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login accounts",
                "parameters": [
                    {
                        "description": "Login user",
                        "name": "tags",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_NghiaLeopard_Go-Ecommerce-Backend_internal_api_handler_request.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_NghiaLeopard_Go-Ecommerce-Backend_internal_api_handler_response.Login"
                        }
                    }
                }
            }
        },
...
        "/api/auth/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "get me",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "get me",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_NghiaLeopard_Go-Ecommerce-Backend_internal_api_handler_response.AuthMe"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update me",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Update me",
                "parameters": [
                    {
                        "description": "Update auth me",
                        "name": "tags",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_NghiaLeopard_Go-Ecommerce-Backend_internal_api_handler_request.UpdateAuthMe"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_NghiaLeopard_Go-Ecommerce-Backend_internal_api_handler_response.UpdateAuthMe"
                        }
                    }
                }
            }
        },
        "/api/auth/refresh-token": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "RefreshToken account",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "RefreshToken accounts",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_NghiaLeopard_Go-Ecommerce-Backend_internal_api_handler_response.GetAccessToken"
                        }
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "description": "Register account",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Register accounts",
                "parameters": [
                    {
                        "description": "Register user",
                        "name": "tags",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_NghiaLeopard_Go-Ecommerce-Backend_internal_api_handler_request.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Register success",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/auth/reset-password": {
            "post": {
                "description": "ResetPassword account",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "ResetPassword accounts",
                "parameters": [
                    {
                        "description": "ResetPassword user",
                        "name": "tags",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_NghiaLeopard_Go-Ecommerce-Backend_internal_api_handler_request.ResetPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reset password success",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_NghiaLeopard_Go-Ecommerce-Backend_db_sqlc.UsersStatus": {
            "type": "string",
            "enum": [
                "0",
                "1"
            ],
            "x-enum-varnames": [
                "UsersStatus0",
                "UsersStatus1"
            ]
        },
        "github_com_NghiaLeopard_Go-Ecommerce-Backend_internal_api_handler_request.ChangePasswordRequest": {
            "type": "object",
            "required": [
                "currentPassword",
                "newPassword"
            ],
            "properties": {
                "currentPassword": {
                    "type": "string",
                    "minLength": 6
                },
                "newPassword": {
                    "type": "string",
                    "minLength": 6
                }
            }
        },
...
        "github_com_NghiaLeopard_Go-Ecommerce-Backend_internal_api_handler_request.ForgotPasswordRequest": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
...
        "github_com_NghiaLeopard_Go-Ecommerce-Backend_internal_api_handler_request.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 6
                }
            }
        },
        "github_com_NghiaLeopard_Go-Ecommerce-Backend_internal_api_handler_request.RegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 6
                }
            }
        },
        "github_com_NghiaLeopard_Go-Ecommerce-Backend_internal_api_handler_request.ResetPasswordRequest": {
            "type": "object",
            "required": [
                "newPassword",
                "secretKey"
            ],
            "properties": {
                "newPassword": {
                    "type": "string",
                    "minLength": 6
                },
                "secretKey": {
                    "type": "string"
                }
            }
        },
...
        "github_com_NghiaLeopard_Go-Ecommerce-Backend_internal_api_handler_request.UpdateAuthMe": {
            "type": "object",
            "required": [
                "address",
                "avatar",
                "city",
                "firstName",
                "lastName",
                "middleName",
                "phoneNumber"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "avatar": {
                    "type": "string"
                },
                "city": {
                    "type": "integer",
                    "format": "int64"
                },
                "firstName": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "middleName": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string",
                    "format": "string"
                }
            }
        },
...
        "github_com_NghiaLeopard_Go-Ecommerce-Backend_internal_api_handler_response.AuthMe": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "integer"
                },
                "address": {
                    "type": "string"
                },
                "addresses": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "avatar": {
                    "type": "string"
                },
                "city": {
                    "type": "integer"
                },
                "create_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "middleName": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/github_com_NghiaLeopard_Go-Ecommerce-Backend_internal_api_handler_response.Role"
                },
                "status": {
                    "$ref": "#/definitions/github_com_NghiaLeopard_Go-Ecommerce-Backend_db_sqlc.UsersStatus"
                }
            }
        },
...
        "github_com_NghiaLeopard_Go-Ecommerce-Backend_internal_api_handler_response.GetAccessToken": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                }
            }
        },
...
        "github_com_NghiaLeopard_Go-Ecommerce-Backend_internal_api_handler_response.Login": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/github_com_NghiaLeopard_Go-Ecommerce-Backend_internal_api_handler_response.User"
                }
            }
        },
...
        "github_com_NghiaLeopard_Go-Ecommerce-Backend_internal_api_handler_response.Role": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "github_com_NghiaLeopard_Go-Ecommerce-Backend_internal_api_handler_response.UpdateAuthMe": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "integer"
                },
                "address": {
                    "type": "string"
                },
                "addresses": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "avatar": {
                    "type": "string"
                },
                "city": {
                    "type": "integer"
                },
                "create_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "middleName": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "integer"
                },
                "role": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/github_com_NghiaLeopard_Go-Ecommerce-Backend_db_sqlc.UsersStatus"
                }
            }
        },
...
        "github_com_NghiaLeopard_Go-Ecommerce-Backend_internal_api_handler_response.User": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "integer"
                },
                "address": {
                    "type": "string"
                },
                "addresses": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "avatar": {
                    "type": "string"
                },
                "city": {
                    "type": "integer"
                },
                "create_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "likeProducts": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "middleName": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "resetToken": {
                    "type": "string"
                },
                "resetTokenExpiration": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/github_com_NghiaLeopard_Go-Ecommerce-Backend_internal_api_handler_response.Role"
                },
                "status": {
                    "$ref": "#/definitions/github_com_NghiaLeopard_Go-Ecommerce-Backend_db_sqlc.UsersStatus"
                },
                "viewedProducts": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Add prefix of Bearer before  token Ex: \"Bearer token\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}